增加裸机版任务管理系统，大彩串口屏显示

+++ Core/Inc/Control.h
@@ -21,39 +21,40 @@ void System_impelment(void);
 
 typedef struct  Control_Struct
 {
-    uint16_t Mscr   : 1;
-    uint16_t Mrelay : 1;
-    uint16_t Sscr   : 1;
-    uint16_t Srelay : 1;
+    uint8_t Mscr   : 1;
+    uint8_t Mrelay : 1;
+    uint8_t Sscr   : 1;
+    uint8_t Srelay : 1;
+    
 }OUTCONTROL;
 
 typedef struct  
 {
-    uint16_t Line_Normal        : 1;
-    uint16_t Line_Abnormal      : 1;
-    uint16_t Line_Malfunction   : 1;
-    uint16_t Line_AB_resume     : 1;
-    uint16_t Line_MAL_resume    : 1;
+    uint8_t Line_Normal        : 1;
+    uint8_t Line_Abnormal      : 1;
+    uint8_t Line_Malfunction   : 1;
+    uint8_t Line_AB_resume     : 1;
+    uint8_t Line_MAL_resume    : 1;
     
 }DEV_EVENT;
 
 typedef struct  
 {
-    uint16_t  Zero_MainCurrent       : 1;
-    uint16_t  Zero_SubCurrent        : 1;
-    uint16_t  Zero_MainVoltage       : 1;
-    uint16_t  Zero_SubVoltage        : 1;
-    uint16_t  Broken_MainPower       : 1;
-    uint16_t  Broken_SubPower        : 1;
+    uint8_t  Zero_MainCurrent       : 1;
+    uint8_t  Zero_SubCurrent        : 1;
+    uint8_t  Zero_MainVoltage       : 1;
+    uint8_t  Zero_SubVoltage        : 1;
+    uint8_t  Broken_MainPower       : 1;
+    uint8_t  Broken_SubPower        : 1;
 }ZERO_STATUS;
 
 
 typedef struct  
 {
-    uint16_t Dev_MainSCR : 1;
-    uint16_t Dev_MainRelay : 1;
-    uint16_t Dev_SubSCR : 1;
-    uint16_t Dev_SubRelay : 1;
+    uint8_t Dev_MainSCR : 1;
+    uint8_t Dev_MainRelay : 1;
+    uint8_t Dev_SubSCR : 1;
+    uint8_t Dev_SubRelay : 1;
 }DEV_STATUE;
 
 // extern DEV_STATUE DEV_CurrentStaute;
